refactor: Move all the VLS-specific changes into the tester project
We'd like to be able to quickly test a new VLS version or a new CLN
version, without having to rebase all of the VLS customizations to the
testing framework in CLN. This is possible because fixture resolution
in `pytest` uses the test module's scope, and not the ones in the
declaration. This means we can override fixtures by redeclaring them,
and then either copying the test, or simpler, just import the test
into the new test module, with different resolution scope, thus the
tests now use a customized `LightningNode` which takes care of
bootstrapping the `lssd` as well as managing a `vlsd2` for each node.